requirements.txt
gradio>=4.41
langchain>=0.2.11
langchain-community>=0.2.10
pypdf>=4.2
sentence-transformers>=3.0
chromadb>=0.5
pandas>=2.2
numpy>=1.26
networkx>=3.3   
python-dotenv>=1.0
groq
